Move blocks to class
